+++ src/main/java/cart/application/CartItemService.java
@@ -4,9 +4,9 @@ import cart.domain.Cart;
 import cart.domain.CartItem;
 import cart.domain.Member;
 import cart.domain.Quantity;
-import cart.dto.CartItemQuantityUpdateRequest;
-import cart.dto.CartItemRequest;
-import cart.dto.CartItemResponse;
+import cart.dto.cartitem.CartItemQuantityUpdateRequest;
+import cart.dto.cartitem.CartItemRequest;
+import cart.dto.cartitem.CartItemResponse;
 import cart.repository.CartItemRepository;
 import cart.repository.ProductRepository;
 import java.util.List;
