+++ src/main/java/cart/dto/cartitem/CartItemQuantityUpdateRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.cartitem;
 
 import javax.validation.constraints.NotNull;
