+++ src/main/java/cart/dto/cartitem/CartItemResponse.java
@@ -1,6 +1,7 @@
-package cart.dto;
+package cart.dto.cartitem;
 
 import cart.domain.CartItem;
+import cart.dto.product.ProductResponse;
 
 public class CartItemResponse {
     private Long id;
