+++ src/main/java/cart/dto/order/OrderRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.order;
 
 import java.util.List;
 import javax.validation.constraints.NotEmpty;
