+++ src/main/java/cart/dto/order/OrderResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.order;
 
 import cart.domain.Order;
 import cart.domain.OrderItem;
