+++ src/main/java/cart/dto/point/MemberPointResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.point;
 
 import cart.domain.Point;
