+++ src/main/java/cart/dto/product/ProductRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.product;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
