+++ src/main/java/cart/ui/PointApiController.java
@@ -3,8 +3,8 @@ package cart.ui;
 import cart.application.PointService;
 import cart.domain.Member;
 import cart.domain.Point;
-import cart.dto.MemberPointResponse;
-import cart.dto.OrderSavedPointResponse;
+import cart.dto.point.MemberPointResponse;
+import cart.dto.order.OrderSavedPointResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
